Replace `switch` with `txs.Visitor` in X chain signer
